fix: route supplier sub category store
