+++ Programming/WebFrontEnd/app/Http/Controllers/Master/SpecializationSupplierController.php
@@ -7,6 +7,7 @@ use App\Http\Controllers\Controller;
 use App\Services\Master\SpecializationSupplier\SpecializationSupplierService;
 use Illuminate\Support\Facades\Session;
 use Illuminate\Support\Facades\Log;
+use App\Http\Requests\Master\Supplier\StoreSupplierSpecialization;
 
 class SpecializationSupplierController extends Controller
 {
@@ -26,8 +27,37 @@ class SpecializationSupplierController extends Controller
     {
     }
 
-    public function store(Request $request)
+    public function store(StoreSupplierSpecialization $request)
     {
+        try {
+            $code = $request->input('supplier_category_name_modal');
+            $name = $request->input('specialization');
+
+            $response = $this->specializationSupplierService->create($code, $name);
+
+            if ($response['metadata']['HTTPStatusCode'] !== 200) {
+                throw new \Exception('Failed to fetch Store Category Supplier => ' . $response['data']['message']);
+            }
+
+            $documentNumbers = array_map(
+                function ($code, $name) {
+                    return "{$code} - {$name}";
+                },
+                $response['data']['subCategoryCode'],
+                $response['data']['subCategoryName']
+            );
+
+            $compact = [
+                "documentNumber" => $documentNumbers,
+                "status" => $response['metadata']['HTTPStatusCode'],
+            ];
+
+            return response()->json($compact);
+        } catch (\Throwable $th) {
+            Log::error("Store Category Supplier Function Error: " . $th->getMessage());
+
+            return response()->json(["status" => 500]);
+        }
     }
 
     public function revision(Request $request)
