+++ Programming/WebFrontEnd/app/Services/Master/SpecializationSupplier/SpecializationSupplierService.php
@@ -46,6 +46,23 @@ class SpecializationSupplierService
         );
     }
 
+    public function create($categoryCode, $specialization)
+    {
+        $token = Session::get('SessionLogin');
+        // $specializationDetails = json_decode($specialization, true);
+
+        return Helper_APICall::setCallAPIGateway(
+            Helper_Environment::getUserSessionID_System(),
+            $token,
+            'transaction.create.master.setSupplierSubCategory',
+            'latest',
+            [
+                "categoryCode" => $categoryCode,
+                "specializationDetails" => $specialization
+            ]
+        );
+    }
+
     public function update($id, $categoryCode, $code, $name, $status)
     {
         $token = Session::get('SessionLogin');
